+++ src/model/moment.ts
@@ -1,11 +1,14 @@
 import { moment as _moment } from "obsidian";
-import type { Moment, MomentInput, MomentFormatSpecification } from "moment";
+import type { Moment, MomentInput, MomentFormatSpecification, Locale } from "moment";
 
-type MomentFactory = (
+type MomentFactory = ((
   input?: MomentInput,
   format?: MomentFormatSpecification,
   strict?: boolean,
-) => Moment;
+) => Moment) & {
+  localeData(): Locale;
+  weekdaysMin(localeSorted?: boolean): string[];
+};
 
 export const moment = _moment as unknown as MomentFactory;
 export type { Moment };